+++ src/xrAICore/Navigation/vertex_path_inline.h
@@ -35,7 +35,7 @@ TEMPLATE_SPECIALIZATION
 inline void CVertexPathBuilder::get_node_path(xr_vector<Index>& path, Vertex* best)
 {
     Vertex *t1 = best, *t2 = best->back();
-    u32 i;
+    size_t i;
     for (i = 1; t2; t1 = t2, t2 = t2->back(), i++)
         ;
     path.resize(i);
